--- helpers/middlewares.js
@@ -1,84 +1,0 @@
-const morgan = require('morgan');
-const loggers = require('./loggers');
-const exceptions = require("./exceptions");
-
-const MAX_PAGE_COUNTER = 100;  // TODO: to make it through env
-let pageCounter = 0;  // TODO: to make it available for other modules
-
-exports.limitContextMiddleware = async function limitContextMiddleware(req, res, next) {
-    if (req.method !== "POST") {
-        next();
-        return;
-    }
-
-    if (pageCounter > MAX_PAGE_COUNTER) {  // TODO: to check for endpoint
-        res.status(429);  // Too many requests
-        res.send({msg: "Page counter expired"});
-        next()
-    }
-
-    next()
-}
-
-/***
- * Returns the middleware for logging HTTP request-response.
-***/
-exports.logHTTPMiddleware = function logHTTPMiddleware() {
-    const logger = loggers.getLogger();
-
-    return morgan(
-        loggers.HTTPFormat,
-        {
-            stream: {
-                write: (message) => logger.http(message),
-            },
-        }
-    );
-}
-
-/***
- * Middleware for processing exceptions.
-***/
-exports.processExceptionMiddleware = async function processExceptionMiddleware(err, req, res, next) {
-    if (res.headersSent) {
-        return next(err);
-    }
-
-    const contextId = err.contextId || req.query.contextId;
-    const pageId = err.pageId || req.query.pageId;
-    const errorMessage = err.message || 'Unknown error';
-
-    if (contextId) {
-        res.header('scrapy-puppeteer-service-context-id', contextId);
-    }
-
-    if (err.contextId) {  // there was a context, but something went wrong
-        res.status(500);
-    } else {  // No context. Possibly, our service was restarted
-        if (err instanceof exceptions.PageNotFoundError || err instanceof exceptions.ContextNotFoundError) {
-            res.status(422);
-        } else {
-            res.status(500);
-        }
-    }
-
-    res.send({
-        contextId,
-        pageId,
-        error: errorMessage
-    });
-
-    next(err);
-}
-
-/***
- * Middleware for logging exceptions.
-***/
-exports.logExceptionMiddleware = async function logExceptionMiddleware(err, req, res, next) {
-    loggers.getLogger().error({
-        message: err,
-        contextId: req.query["contextId"],
-        pageId: req.query["pageId"],
-    });
-    next();
-}
